feat: Core Value 모바일 생성
