+++ apps/web/src/components/mainpage/coreValue/ValueCard.tsx
@@ -1,5 +1,5 @@
 import clsx from 'clsx';
-import { MotionValue, useTransform } from 'framer-motion';
+import { MotionValue } from 'framer-motion';
 import { Orbitron } from 'next/font/google';
 import { FC, ReactNode } from 'react';
 
@@ -15,39 +15,29 @@ interface ValueCardProps {
   name: string;
   description: ReactNode;
   seq: number;
-  centerLineProgress: MotionValue<number>;
-  flipRange: [number, number];
+  flipValue: MotionValue<number>;
 }
 
-const ValueCard: FC<ValueCardProps> = ({
-  className,
-  image,
-  keyword,
-  name,
-  description,
-  seq,
-  flipRange,
-  centerLineProgress,
-}) => {
-  const flip = useTransform(centerLineProgress, flipRange, [0, 1]);
-
+const ValueCard: FC<ValueCardProps> = ({ className, image, keyword, name, description, seq, flipValue }) => {
   const formattedSeq = `${seq}`.padStart(2, '0');
 
   return (
     <FlipableCard
       className={className}
-      flipValue={flip}
+      flipValue={flipValue}
       front={
         <div className='bg-gray2 relative flex h-full flex-col overflow-clip rounded-[0.8rem]'>
-          <div className=''>{image}</div>
+          <div className='[&>svg]:h-full [&>svg]:w-full'>{image}</div>
           <div className='flex flex-1 flex-col items-center justify-center'>
-            <span className={clsx(orbitron.className, 'text-[3.6rem] font-bold leading-[100%]')}>{keyword}</span>
-            <span className='mt-[2rem] text-[2.8rem] font-semibold'>{name}</span>
+            <span className={clsx(orbitron.className, 'text-[1.4rem] font-bold leading-[100%] md:text-[3.6rem]')}>
+              {keyword}
+            </span>
+            <span className='mt-[0.76rem] text-[1.2rem] font-semibold md:mt-[2rem] md:text-[2.8rem]'>{name}</span>
           </div>
           <div
             className={clsx(
               orbitron.className,
-              'absolute left-0 right-0 top-[3.2rem] flex justify-center text-[2.4rem]',
+              'absolute left-0 right-0 top-[1.18rem] flex justify-center text-[1rem] md:top-[3.2rem] md:text-[2.4rem]',
             )}
           >
             {formattedSeq}
@@ -56,11 +46,11 @@ const ValueCard: FC<ValueCardProps> = ({
       }
       back={
         <div className='relative h-full overflow-clip rounded-[0.8rem]'>
-          <div className='text-24-semibold h-full w-full'>{description}</div>
+          <div className='md:text-24-semibold h-full w-full text-[1rem]'>{description}</div>
           <div
             className={clsx(
               orbitron.className,
-              'absolute left-0 right-0 top-[3.2rem] flex justify-center text-[2.4rem]',
+              'absolute left-0 right-0 top-[1.18rem] flex justify-center text-[1rem] md:top-[3.2rem] md:text-[2.4rem]',
             )}
           >
             {formattedSeq}
